Code to get release information:
Wrote code in test_release.py to pipe commands through the shell (namely 'lsb_release -a'), and return the output to a variable.

I wanted to be able to run the lsb_release command and extract information such as Linux distro, release number, etc. from the output, but I was having difficulty working with the output directly. Instead, the program now writes the output to a text file, and I have a function that runs through the lines of the output and extracts what I'm looking for, which works well.

Next I intend to create a JSON file populated with release information, and use that to build a class "this_pc," which will have instructions on commands to use for the current PC and its OS.

This has only been tested on a single Ubuntu computer.